🔨 [TEA-68]: Restructure router from main.go into new pkg for unit testing
